First Merge
I opened my project and then exited out because I realized that I needed to merge to the main so I could use pressure plates and such. GitHub then said the.vsconfig needed to go so I think it wants me to commit this first. I apologize if that is not what I was supposed to do.